requirements.txt
aiogram==3.25.0
aiosqlite==0.22.1
python-dotenv==1.2.1

# PyTorch (CPU). Для GPU замени на нужную версию: https://pytorch.org/get-started/locally/
torch==2.6.0
torchvision==0.21.0

# Обучение
scikit-learn==1.6.1
matplotlib==3.10.1
seaborn==0.13.2
tensorboard==2.19.0

numpy==2.2.3
Pillow==11.1.0